Remove PHP code, clean up code
I removed unnecessary PHP code, converting to year code in the footer to JavaScript. This makes the site easier to port if we need to move away from PHP.
I also cleaned up a lot of the HTML tags, fixed some tag mismatches, and added a link to my site in tiny text in the footer.